+++ dcpy/lifecycle/ingest/templates/dcp_atomicpolygons.yml
