+++ dcpy/lifecycle/ingest/templates/dcp_pad.yml
