+++ dcpy/lifecycle/ingest/templates/nysdoh_nursinghomes.yml
