+++ dcpy/test/lifecycle/ingest/__init__.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+RESOURCES = Path(__file__).parent / "resources"
+TEST_DATA_DIR = "test_data"
+FAKE_VERSION = "v001"
