+++ dcpy/test/lifecycle/ingest/resources/tempate.yml
