+++ dcpy/test/lifecycle/ingest/resources/test_data/test.gdb/a00000003.gdbtable
