+++ dcpy/test/lifecycle/ingest/resources/test_data/test.gdb/a00000004.gdbtablx
