+++ dcpy/test/lifecycle/ingest/resources/test_data/test.gdb/a00000006.gdbtablx
