+++ dcpy/test/lifecycle/ingest/resources/test_data/test.gdb/a0000000b.spx
